Restrict locale settings to the file they occur in
This prevents LANG= in an included file from affecting the interpretation
of month and day names in the including file.

Make the internal pre-processor accept white space between the "#" at
the start of the line and the keyword for better compatibility with cpp.

Add support for the cpp keywords #warning and #error.

MFC after:	3 days


git-svn-id: https://svn.freebsd.org/base/head@367364 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f