MFC 365842,365844: Remove support for setting some obscure fields.
365842:
Remove support for setting some obscure fields.

Don't permit setting the exception bitmap or VMCS entry interrupt
information.  These are not generally useful to set.  If it is needed
in the future, dedicated pseudo registers can be added for these that
would be used with vm_set_register().

365844:
Remove no longer used variable.


git-svn-id: https://svn.freebsd.org/base/stable/12@367408 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f