rms: several cleanups + debug read lockers handling
This adds a dedicated counter updated with atomics when INVARIANTS
is used. As a side effect one can reliably determine the lock is held
for reading by at least one thread, but it's still not possible to
find out whether curthread has the lock in said mode.

This should be good enough in practice.

Problem spotted by avg.


git-svn-id: https://svn.freebsd.org/base/head@367453 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f