MFC r367269: Fix for referencing file via its vnode in ibore.
Use the native vnode lookup functions, instead of going via the LinuxKPI,
because the file referenced is typically created outside the LinuxKPI, and
the LinuxKPI's fdget() can only resolve file descriptor numbers which
were created by itself.

The vnode pointer is used as an identifier to identify XRCD handles which
are sharing resources.

This patch fixes the so-called XRCD support in ibcore for FreeBSD.
Refer to ibv_open_xrcd(3) for more information how the file descriptor
argument is used.

Reviewed by:	kib@
Sponsored by:	Mellanox Technologies // NVIDIA Networking


git-svn-id: https://svn.freebsd.org/base/stable/12@367562 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f