MFC r366430: ngctl: add -c (compact output) for the dot command
The output of "ngctl dot" is suitable for small netgraph networks. Even
moderate complex netgraph setups (about a dozen nodes) are hard to
understand from the .dot output, because each node and each hook are shown
as a full blown structure.

This patch allows to generate much more compact output and graphs by
omitting the extra structures for the individual hooks. Instead the names of
the hooks are labels to the edges.


git-svn-id: https://svn.freebsd.org/base/stable/12@367664 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f