ena-com: Fix ena-com to allocate cdesc aligned to 4k
The latest generation hardware requires IO CQ (completion queue)
descriptors memory to be aligned to a 4K. It needs that feature for
the best performance.

Allocating unaligned descriptors will have a big performance impact as
the packet processing in a HW won't be optimized properly.

It's a critical fix, especially for the arm64 EC2 instances.


git-svn-id: https://svn.freebsd.org/base/vendor-sys/ena-com/dist@367793 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f